requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg[binary]==3.2.12
openai>=1.30.0
python-dotenv>=1.1.0
bcrypt>=4.0.0
email-validator>=2.0.0